requirements.txt
streamlit==1.43.2
tensorflow==2.21.0
numpy==1.26.4
pillow==11.1.0
